separating out wrap tests
