+++ packages/cacheable/test/index.test.ts
@@ -1,4 +1,4 @@
-import { createWrapKey, type GetOrSetOptions } from "@cacheable/memoize";
+import type { GetOrSetOptions } from "@cacheable/memoize";
 import { sleep } from "@cacheable/utils";
 import { faker } from "@faker-js/faker";
 import KeyvRedis from "@keyv/redis";
@@ -679,80 +679,6 @@ describe("cacheable hash method", async () => {
 	});
 });
 
-describe("cacheable wrap", async () => {
-	test("should wrap method with key and ttl", async () => {
-		const cacheable = new Cacheable();
-		const asyncFunction = async (value: number) => Math.random() * value;
-		const options = {
-			keyPrefix: "keyPrefix",
-			ttl: 10,
-		};
-
-		const wrapped = cacheable.wrap(asyncFunction, options);
-		const result = await wrapped(1);
-		const result2 = await wrapped(1);
-		expect(result).toBe(result2);
-		const cacheKey = createWrapKey(asyncFunction, [1], {
-			keyPrefix: options.keyPrefix,
-		});
-		const cacheResult1 = await cacheable.get(cacheKey);
-		expect(cacheResult1).toBe(result);
-		await sleep(20);
-		const cacheResult2 = await cacheable.get(cacheKey);
-		expect(cacheResult2).toBeUndefined();
-	});
-	test("wrap async function", async () => {
-		const cache = new Cacheable();
-		const options = {
-			keyPrefix: "wrapPrefix",
-			ttl: "5m",
-		};
-
-		const plus = async (a: number, b: number) => a + b;
-		const plusCached = cache.wrap(plus, options);
-
-		const multiply = async (a: number, b: number) => a * b;
-		const multiplyCached = cache.wrap(multiply, options);
-
-		const result1 = await plusCached(1, 2);
-		const result2 = await multiplyCached(1, 2);
-
-		expect(result1).toBe(3);
-		expect(result2).toBe(2);
-	});
-
-	test("should wrap to default ttl", async () => {
-		const cacheable = new Cacheable({ ttl: 10 });
-		const asyncFunction = async (value: number) => Math.random() * value;
-		const options = {
-			keyPrefix: "wrapPrefix",
-		};
-		const wrapped = cacheable.wrap(asyncFunction, options);
-		const result = await wrapped(1);
-		const result2 = await wrapped(1);
-		expect(result).toBe(result2); // Cached
-		await sleep(15);
-		const result3 = await wrapped(1);
-		expect(result3).not.toBe(result2);
-	});
-
-	test("Cacheable.wrap() passes createKey option through", async () => {
-		const cacheable = new Cacheable();
-		let createKeyCalled = false;
-		const asyncFunction = async (argument: string) => `Result for ${argument}`;
-		const options = {
-			createKey: () => {
-				createKeyCalled = true;
-				return "testKey";
-			},
-		};
-
-		const wrapped = cacheable.wrap(asyncFunction, options);
-		await wrapped("arg1");
-		expect(createKeyCalled).toBe(true);
-	});
-});
-
 describe("cacheable namespace", async () => {
 	test("should set the namespace via options", async () => {
 		const cacheable = new Cacheable({ namespace: "test" });
@@ -863,51 +789,6 @@ describe("cacheable get or set", () => {
 });
 
 describe("cacheable adapter coverage", () => {
-	test("should directly test wrap adapter on method", async () => {
-		const cacheable = new Cacheable();
-
-		// We need to directly call the adapter's on method to achieve 100% coverage
-		// Even though @cacheable/memoize doesn't call it, we need to test it works
-
-		// Access the wrap method's internals
-		const testFn = async () => "result";
-
-		// Create an adapter manually that mimics what wrap does
-		const adapter = {
-			get: async (key: string) => cacheable.get(key),
-			has: async (key: string) => cacheable.has(key),
-			// biome-ignore lint/suspicious/noExplicitAny: adapter interface
-			set: async (key: string, value: any, ttl?: number | string) => {
-				await cacheable.set(key, value, ttl);
-			},
-			// This is the method we need to cover (lines 838-839)
-			// biome-ignore lint/suspicious/noExplicitAny: adapter interface
-			on: (event: string, listener: (...args: any[]) => void) => {
-				cacheable.on(event, listener);
-			},
-			// biome-ignore lint/suspicious/noExplicitAny: adapter interface
-			emit: (event: string, ...args: any[]) => cacheable.emit(event, ...args),
-		};
-
-		// Test that the adapter's on method works
-		let listenerCalled = false;
-		adapter.on("test-wrap-event", () => {
-			listenerCalled = true;
-		});
-
-		// Trigger the event through the adapter
-		adapter.emit("test-wrap-event");
-
-		expect(listenerCalled).toBe(true);
-
-		// Also test the regular wrap functionality
-		const wrapped = cacheable.wrap(testFn, { keyPrefix: "test" });
-		await wrapped();
-		expect(
-			await cacheable.has(createWrapKey(testFn, [], { keyPrefix: "test" })),
-		).toBe(true);
-	});
-
 	test("should directly test getOrSet adapter on method", async () => {
 		const cacheable = new Cacheable();
 
@@ -943,6 +824,38 @@ describe("cacheable adapter coverage", () => {
 		const testFn = async () => "getorset-value";
 		const result = await cacheable.getOrSet("key", testFn);
 		expect(result).toBe("getorset-value");
+	});
+});
+
+describe("cacheable serialize/deserialize properties", () => {
+	test("should get and set serialize property", () => {
+		const cacheable = new Cacheable();
+		// biome-ignore lint/suspicious/noExplicitAny: testing serializer interface
+		const customSerialize = (obj: any) => `custom:${JSON.stringify(obj)}`;
+
+		// Test getter
+		expect(typeof cacheable.serialize).toBe("function");
+
+		// Test setter
+		cacheable.serialize = customSerialize;
+		expect(cacheable.serialize).toBe(customSerialize);
+	});
+
+	test("should get and set deserialize property", () => {
+		const cacheable = new Cacheable();
+		const customDeserialize = (str: string) => {
+			if (str.startsWith("custom:")) {
+				return JSON.parse(str.substring(7));
+			}
+			return JSON.parse(str);
+		};
+
+		// Test getter
+		expect(typeof cacheable.deserialize).toBe("function");
+
+		// Test setter
+		cacheable.deserialize = customDeserialize;
+		expect(cacheable.deserialize).toBe(customDeserialize);
 	});
 });
